refactor: 清理模块直接返回收尾
按 RET504 清理剩余模块中赋值后立刻返回的临时变量，覆盖 combat、commission、config、dorm、event、exercise、handler、raid、research、retire、statistics。

修正自动修复留下的注释位置和空行，并中文化触碰到的说明。

+++ module/handler/info_handler.py
@@ -25,8 +25,7 @@ def info_letter_preprocess(image):
     image = (image - 64) / 0.75
     image[image > 255] = 255
     image[image < 0] = 0
-    image = image.astype("uint8")
-    return image
+    return image.astype("uint8")
 
 
 class InfoHandler(ModuleBase):
